implemented interfaces for stats gathering

+++ Assets/Scripts/Interfaces/IClientStats.cs
@@ -5,12 +5,12 @@ using System.Collections.Generic;
 public interface IClientStats
 {
     // присылает активную мощность в каунтер
-    event Action<string, int> OnGoingOffline;
+    event Action<string, float> OnGoingOffline;
 
-    event Action<string, int> OnGoingOnline;
+    event Action<string, float> OnGoingOnline;
 
     // присылает полезную мощность в каунтер
-    event Action<string, int> OnBusyMode;
+    event Action<string, float> OnBusyMode;
 
-    event Action<string, int> OnIdleMode;
+    event Action<string, float> OnIdleMode;
 }
